Glass restructure, still not right in PathTracing

+++ renderer.h
@@ -10,7 +10,7 @@ namespace Tmpl8
 		void Init();
 		float3 Trace(Ray& ray, int iter);
 		float3 PathTrace(Ray& ray, int n);
-		float3 Path(Ray& ray, int iter);
+		float3 Path(Ray& ray, float iter);
 		float DirectIllumination(float3& I);
 		float3 Absorb(float3 color, float d, float3 absorpMat);
 		void Tick(float deltaTime);
